+++ chrome/CHANGES.md
@@ -7,6 +7,8 @@ CHANGE LOG:
 Version 2.5.0.001
 -------------------
 - Add support to attach images from Gmail
+- Fix typo with missing brace for uriForDisplay
+- To handle jQuery UI looking to Gmail for UI icons, must replace url("images... with url("chrome-extension://__MSG_@@extension_id__/images...
 
 Version 2.4.27
 ---------------